move handleDownload functionality to helper.ts file
